+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [0.15.4] - 2021-08-19
+
+## Fixes
+
+-   Fixes extra semicolon at the end of thirdparty sign in up component.
+
 ## [0.15.3] - 2021-08-04
 
 ## Fixes